Extract recovery handlers to dedicated module (Phase 3)
Created src/lib/recovery-handlers.ts to house terminal session recovery functions:
- handleRecoverNewSession(): Create new session for terminal with missing session
- handleRecoverAttachSession(): Show list of available sessions for attachment
- handleAttachToSession(): Attach terminal to existing session
- handleKillSession(): Kill orphaned session and refresh list

These functions use dependency injection pattern to receive:
- state (AppState)
- generateNextConfigId function
- saveCurrentConfig function

Benefits:
- Reduced main.ts from 2,117 to 1,983 lines (134 lines removed)
- Recovery logic now in focused module (213 lines)
- Clear separation of concerns
- Easier to test and maintain

Phase 3 continues the incremental refactoring strategy from issue #237.
